Add separate command for symbol on background sheet
The background sheet has an option for a character's organization
name but not for an associated picture. Also, when one would add
a picture without text, it would overwrite the organization name.
Make things easier and add a separate command for
\OrganizationSymbol .

Added an example for Aaliyah, who just became a member of the
Shadow Thieves!