Separate symbol command (#119)
* Couple of spelling fixes

* Add etoolbox package

This fixes compilation without the dnd latex style. Otherwise
compile fails because \ifdimequal is not defined. In particular,
this fixes compilation of the aaliyah-plain.tex example

* Add separate command for symbol on background sheet

The background sheet has an option for a character's organization
name but not for an associated picture. Also, when one would add
a picture without text, it would overwrite the organization name.
Make things easier and add a separate command for
\OrganizationSymbol .

Added an example for Aaliyah, who just became a member of the
Shadow Thieves!